+++ src/NewButton/button-base.tsx
@@ -1,0 +1,6 @@
+import styled from 'styled-components'
+import sx, {SxProp} from '../sx'
+
+const ButtonBase = styled.button<SxProp>(sx)
+
+export default ButtonBase
